refactor: split audiosdk
Split audiosdk into AudioRecorder and AudioPlayer. The original file
was nearly 800 lines and thus too long.
